// frontend/teampilot/src/app/services/api/v1/model/tournamentDTO.ts
/**
 * TeamPilot.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PastMatchDTO } from './pastMatchDTO';
import { MatchDTO } from './matchDTO';
import { TeamForTournamentDTO } from './teamForTournamentDTO';


export interface TournamentDTO {
    tournamentId?: string | null;
    tournamentName?: string | null;
    tournamentFormat?: string | null;
    tournamentPrizePool?: number;
    tournamentLocation?: string | null;
    tournamentSponsor?: string | null;
    tournamentStartDate?: string | null;
    tournamentEndDate?: string | null;
    tournamentMatches?: Array<MatchDTO> | null;
    futureTournamentMatches?: Array<MatchDTO> | null;
    pastTournamentMatches?: Array<PastMatchDTO> | null;
    tournamentParticipatingTeams?: Array<TeamForTournamentDTO> | null;
    winningTeam?: TeamForTournamentDTO;
    upcomingMatch?: MatchDTO;
}